Rename demo app package name.
