+++ PopupBridgeExample/src/androidTest/java/com/braintreepayments/popupbridge/demo/test/PopupBridgeTest.java
@@ -1,4 +1,4 @@
-package com.braintreepayments.api.example.test;
+package com.braintreepayments.popupbridge.demo.test;
 
 import android.os.Build;
